move and refactor getUpdatedHoverHeader to reduce circular dependencies
